pyproject.toml
[project]
name = "aimathagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "python-dotenv>=1.2.1",
]
